+++ test/clj_surgeon/help_test.clj
@@ -102,7 +102,7 @@
                      :rename-ns :rename-ns!
                      :fix-declares :fix-declares!
                      :extract :extract!
-                     :find-subform :lens :edit :replace-subform :replace-subform!
+                     :find-subform :lens :xray :edit :replace-subform :replace-subform!
                      :cljc-merge :cljc-split :cljc-add-require :cljc-analyze}]
       (is (= expected (set (keys core/ops-registry)))))))